% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_biplot}
\alias{nano_biplot}
\title{Bivariate plot of length against base call quality
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2}
\usage{
nano_biplot(inputfile, viridiscolor = "viridis")
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}
}
\value{
Bivariate plot of read lengths vs read quality
}
\description{
Bivariate plot of length against base call quality
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2
}
