% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_channel}
\alias{nano_channel}
\title{Heatmap of reads per channel
Takes in the `sequencing_summary` file of albacore as input and visualizes the reads generated per channel over time}
\usage{
nano_channel(inputfile, viridiscolor = "viridis")
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}
}
\value{
heatmap of reads per channel
}
\description{
Heatmap of reads per channel
Takes in the `sequencing_summary` file of albacore as input and visualizes the reads generated per channel over time
}
