% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{nano_histlength}
\alias{nano_histlength}
\title{Histogram of read lengths of nanopore reads
Takes in the `sequencing_summary` file of albacore as input and visualizes the readlength distribution as a histogram using ggplot2}
\usage{
nano_histlength(inputfile, viridiscolor = "viridis", transformation = "")
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}

\item{transformation}{"log10" transformation or leave it empty for regular plot}
}
\value{
Histogram of read lengths
}
\description{
Histogram of read lengths of nanopore reads
Takes in the `sequencing_summary` file of albacore as input and visualizes the readlength distribution as a histogram using ggplot2
}
