% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_joy_length}
\alias{nano_joy_length}
\title{joy distribution of read length
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2}
\usage{
nano_joy_length(inputfile, viridiscolor = "viridis", transformation = "")
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}

\item{transformation}{"log10" transformation or leave it empty for regular plot}
}
\value{
joy distribution of read length
}
\description{
joy distribution of read length
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2
}
