% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{nano_joy_quality}
\alias{nano_joy_quality}
\title{joy distribution of read quality
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2}
\usage{
nano_joy_quality(inputfile, viridiscolor = "viridis")
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}
}
\value{
joy distribution of read quality
}
\description{
joy distribution of read quality
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2
}
