% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_longest}
\alias{nano_longest}
\title{Report n longest reads of a sequencing run
Takes in the `sequencing_summary` file of albacore as input and give you the 'n' longest sequences}
\usage{
nano_longest(inputfile, quality = 0, n = 5)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{quality}{Mean quality filter option for every read}

\item{n}{longest reads}
}
\value{
basic nanopore run statistics
}
\description{
Report n longest reads of a sequencing run
Takes in the `sequencing_summary` file of albacore as input and give you the 'n' longest sequences
}
