% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_n50}
\alias{nano_n50}
\title{N50 of nanopore run
Takes in the `sequencing_summary` file of albacore as input and returns the N50}
\usage{
nano_n50(inputfile, quality = 0)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{quality}{Mean quality filter option for every read}
}
\value{
N50
}
\description{
N50 of nanopore run
Takes in the `sequencing_summary` file of albacore as input and returns the N50
}
