% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{nano_readsum}
\alias{nano_readsum}
\title{Summary of n longest reads and their basecall quality
Takes in the `sequencing_summary` file of albacore as input and give you a summary of longest reads and their quality}
\usage{
nano_readsum(inputfile, number = 5)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}
}
\value{
summary of n longest reads
}
\description{
Summary of n longest reads and their basecall quality
Takes in the `sequencing_summary` file of albacore as input and give you a summary of longest reads and their quality
}
