% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_seq_cov}
\alias{nano_seq_cov}
\title{Genome coverage of nanopore run
Takes in the `sequencing_summary` file of albacore and the genome fasta as input and returns the genome coverage}
\usage{
nano_seq_cov(inputfile, genome_fasta, quality = 0)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{genome_fasta}{path to genome fasta}

\item{quality}{Mean quality filter option for every read}
}
\value{
Genome coverage = Nr of bases sequenced/genome length
}
\description{
Genome coverage of nanopore run
Takes in the `sequencing_summary` file of albacore and the genome fasta as input and returns the genome coverage
}
