% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_seq_vs_exon}
\alias{nano_seq_vs_exon}
\title{Read length direct RNA seq vs expected transcript length
Takes in the `sequencing_summary` file of albacore as input and a gff3 annotation file and returns a comparison of read length vs transcript length}
\usage{
nano_seq_vs_exon(inputfile, gff_file, quality = 0)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{gff_file}{path to gff3 annotation file}

\item{quality}{Mean quality filter option for every read}
}
\value{
comparison of read_length vs transcript length
}
\description{
Read length direct RNA seq vs expected transcript length
Takes in the `sequencing_summary` file of albacore as input and a gff3 annotation file and returns a comparison of read length vs transcript length
}
