% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{nano_stats}
\alias{nano_stats}
\title{Summary of basic nanopore run statistics
Takes in the `sequencing_summary` file of albacore as input and give you a basic summary}
\usage{
nano_stats(inputfile)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}
}
\value{
joy distribution of read quality
}
\description{
Summary of basic nanopore run statistics
Takes in the `sequencing_summary` file of albacore as input and give you a basic summary
}
