% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_yield}
\alias{nano_yield}
\title{Cumulative yield plot
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2}
\usage{
nano_yield(inputfile, viridiscolor = "viridis")
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{viridiscolor}{Viridis color palette option}
}
\value{
Cumulative yield plot
}
\description{
Cumulative yield plot
Takes in the `sequencing_summary` file of albacore as input and visualizes the readquality distribution as a histogram using ggplot2
}
