% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoRaw_scripts.R
\name{nano_stats}
\alias{nano_stats}
\title{Summary of basic nanopore run statistics
Takes in the `sequencing_summary` file of albacore as input and give you a basic summary}
\usage{
nano_stats(inputfile, quality = 0)
}
\arguments{
\item{inputfile}{Filepath to sequencing_summary.txt output of albacore}

\item{quality}{Mean quality filter option for every read}
}
\value{
basic nanopore run statistics
}
\description{
Summary of basic nanopore run statistics
Takes in the `sequencing_summary` file of albacore as input and give you a basic summary
}
